MINOR: cli: display failure reason on wait command
wait CLI command can be used to wait until a specific condition is
reached. So far, srv-removable is the only event supported. This is
implemented via srv_check_for_deletion(), which is able to report a
message describing the reason if the condition is unmet.

Previously, wait handler did not reuse the provided message. Instead,
generic errors were reported, which might not be enough for debugging.
Improve this so that now wait is able to report the final message when
it returns, either on a fatal error or on timeout. This requires the
usage of dynamic string output using cli_dynmsg/err functions.