MEDIUM: mux-quic: protect against idle conn usage
Add protection in QUIC MUX when an idle connection is being used. Even
if takeover is not yet implemented, this is useful as idle connections
can be purged by any thread, for example when a server goes into
maintenance.

This change should prevent crashes related to idle connections when a
QUIC server is removed via "delete server".