MINOR: connection: remove extra session_unown_conn() on reverse
When a connection is reversed via rhttp protocol, on the edge side it
will migrate from the frontend to the backend side. This is performed
via conn_reverse(). During this operation, the owning session is removed
as it becomes unneeded.

Special care must be taken first. <origin> member of session is set to
NULL. This ensures that the session won't delete the reversed
connection. Prior to this patch, session_unown_conn() was also called.
However, this is not necessary as this is only used to detach a backend
connection from a session in the context of reuse.

As such, this patch removes session_unown_conn() invokation during
conn_reverse().