MEDIUM: session: reject idle conns if server in maintenance
Previous patch ensures that idle connections are now rejected if server
is in maintenance. This patch introduces a similar change for
connections stored in the session instead.

Via session_check_idle_conn(), if connection target server is in
maintenance, conn is removed from the session and an error is returned,
similarly to session max idle conns limit reached.