BUG/MINOR: acl: Fix error message about several '-m' parameters
There is a typo in the commit * c51ddd5c3 ("MINOR: acl: Only allow one '-m'
matching method") . '*m' was reported in the error message instead of '-m'.

In addition, it is now mentionned that only the last one should be keep if
an old config triggers the error.

No backport needed, except if the commit above is backported.